requirements.txt
pip==24.0.0
PyYAML==6.0.1
setuptools==69.1.0
